+++ NEWS.md
@@ -1,4 +1,5 @@
-# Ternary v1.0.2.9002 (development)
+# Ternary v1.0.2.9100 (development)
+ - Add support for contour plots
  - Maintain isometry when one of `xlim` or `ylim` is specified.
  - Fix issues with `TernaryPlot` documentation.
  - Remove names in `TernaryCoords`.